requirements.txt
absl-py==1.0.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
boto3==1.20.23
botocore==1.23.23
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
cycler==0.11.0
fonttools==4.28.5
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
h11==0.12.0
idna==3.3
importlib-metadata==4.10.0
jmespath==0.10.0
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.5.62
outcome==1.1.0
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
plotly==5.5.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.8
s3transfer==0.5.0
scipy==1.7.3
seaborn==0.11.2
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
zipp==3.6.0
